Refactor task package name.
